added setResume to IRequest, removed some debug output
git-svn-id: https://svn.intern.artcom.de/pro60/trunk/src@2839 f3027e6b-e3f5-0310-a8e1-edde99107618

+++ Y60/acgtk/DragButton.h
@@ -39,8 +39,8 @@ class DragButton : public EmbeddedButton {
         bool on_leave_notify_event(GdkEventCrossing * theEvent);
     private:
         bool _myDragInProgressFlag;
-        double _myLastX;
-        double _myLastY;
+        int _myLastX;
+        int _myLastY;
     
         sigc::signal<void, double, double> _myDragStartSignal;
         sigc::signal<void, double, double> _myDragSignal;
